Add files via upload
Why is the mortality rate higher in developed countries? It is clear that the Coronavirus pandemic will remain on the world agenda for a few more months. When we look at the information shared by different institutions, organizations, or experts, it can be seen that we have not yet reached satisfactory information about the pandemic and its spread. However, it is accepted that all kinds of contact reduction and isolation are effective in terms of protection. In this project we analyzed government data, healthcare data and data shared by Dr. Anupam Sule from St Joseph Mercy Hospital, to understand the spread of virus in all counties of USA. Not only this report has trends, pandemic variables calculated like waves, severity, and magnitude of pandemic but it also allows to draw a meaningful insight for the government and health care system of United States America to formulate policies and measure to control the spread of the virus. The counties are distinguished on Rural Urban Continuum Code and we analyzed four counties where first county is most urban ones followed by second, third and fourth. My analysis is focused on County 1and County 4.